requirements.txt
Django==4.2.1
django-cors-headers==3.14.0
django-filter==22.1
django-jazzmin==2.6.0
django-sslserver==0.22
django_filter==22.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
python_dateutil==2.8.2
requests==2.29.0
rest_framework_simplejwt==0.0.2
requests==2.29.0
flake8==6.0.0
psycopg2-binary==2.9.6
Pillow==9.5.0
python-dotenv==1.0.0